pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-pipeline"
version = "0.1.0"
description = "Modular video data processing + vLLM offline inference pipeline (multimodal / Qwen-VL video)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Your Team" }]
dependencies = [
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "transformers>=4.56",
  "accelerate>=0.34",
  "qwen-vl-utils",
  "vllm>=0.11.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
  "pytest>=8.0",
]

[project.scripts]
video-pipeline-run = "video_pipeline.cli.run:main"

[tool.setuptools]
packages = ["video_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
